Operational Helm Chart & Container Image
Originally had a functional container image, but ran into issues with
nodes and the inform endpoint. Found in server.log there was missing
shared libraries expected (systemd, etc) so moved to the Ubuntu base
image and pulled in dependencies to make it happy (UniFi java
application).

This current helm chart is deployed and being used; however, still need
to test out additional functionality as well as migrate to using
certificates for HTTPS endpoints.