Helm Chart Adjustments; Image Entry Changes; Updated Readme
The entrypoint can now take a TLS certificate and load it into the
keystore used by UniFi; however, will require restarting the pod in the
event the cert rolls. Added the option to place a keystore directly in
and began testing with cert-manager; however, cert-manager isnt
currently able to set an alias until the June release. The Gateway files
I need to go back over and template for use with the helm chart, but in
the meantime placing it as an example for what I have deployed.

Updated readme